test: move tests to own folder
